+++ axle-wheel/src/test/scala/axle/stats/BayesTheoremSpec.scala
@@ -7,24 +7,13 @@ import spire.math.Rational
 
 import axle.eqSymbol
 import axle.algebra.Region
-import axle.algebra.RegionEq
-
-object TestSupport {
-
-  val arbitraryPortion: Arbitrary[Rational] =
-    Arbitrary(for {
-      denominator <- Gen.oneOf(1 to 1000)
-      numerator <- Gen.oneOf(0 to denominator)
-    } yield Rational(numerator.toLong, denominator.toLong))  
-
-}
 
 class FairCoinIsBayes
   extends BayesTheoremProperty[Unit, ConditionalProbabilityTable, Symbol, Rational](
     "Fair coin",
     Arbitrary(Gen.oneOf(List(()))),
     u => coin(),
-    m => Arbitrary(Gen.oneOf(coinSides.map(RegionEq(_)))),
+    m => Arbitrary(TestSupport.genRegion(coinSides)),
     m => Region.eqRegionIterable(coinSides))
 
 import spire.math.Rational
@@ -33,7 +22,7 @@ class BiasedCoinIsBayes
     "Arbitrarily biased coins",
     Arbitrary(Gen.choose(0d,1d).map(Rational.apply)),
     coin,
-    bias => Arbitrary(Gen.oneOf(coinSides.map(RegionEq(_)))),
+    bias => Arbitrary(TestSupport.genRegion(coinSides)),
     bias => Region.eqRegionIterable(coinSides))
 
 import spire.implicits.IntAlgebra
@@ -44,7 +33,7 @@ class D6IsBayes
     "dice",
     Arbitrary(Gen.oneOf(List(4,6,8,10,12,20))),
     die,
-    n => Arbitrary(Gen.oneOf((1 to n).map(RegionEq(_)))), // TODO random expression
+    n => Arbitrary(TestSupport.genRegion(1 to n)),
     n => Region.eqRegionIterable(1 to n))
 
 class TwoPlatonicSolidDieAddedBayes
@@ -60,7 +49,7 @@ class TwoPlatonicSolidDieAddedBayes
            a + b           
        }}
     },
-    { case (an, bn) => Arbitrary(Gen.oneOf((1 to an*bn).map(RegionEq(_)))) }, // TODO random expression
+    { case (an, bn) => Arbitrary(TestSupport.genRegion(1 to an*bn)) },
     { case (an, bn) => Region.eqRegionIterable(1 to an*bn) }
 )
 
@@ -76,15 +65,9 @@ class AlarmBurglaryEarthquakeBayesianNetworkIsBayes
     (Boolean, Boolean, Boolean, Boolean, Boolean),
     Rational](
     "Alarm-Burglary-Earthquake Bayesian Network",
-    TestSupport.arbitraryPortion,
-    { case seed => MonotypeBayesanNetwork(
-        new AlarmBurglaryEarthquakeBayesianNetwork(pEarthquake = seed).bn,
-        AlarmBurglaryEarthquakeBayesianNetwork.select,
-        AlarmBurglaryEarthquakeBayesianNetwork.combine1,
-        AlarmBurglaryEarthquakeBayesianNetwork.combine2)
-    },
-    // TODO random expression
-    { case seed => Arbitrary(Gen.oneOf(AlarmBurglaryEarthquakeBayesianNetwork.domain.map(RegionEq(_)))) },
+    Arbitrary(TestSupport.genPortion),
+    { case seed => new AlarmBurglaryEarthquakeBayesianNetwork(pEarthquake = seed).monotype },
+    { case seed => Arbitrary(TestSupport.genRegion(AlarmBurglaryEarthquakeBayesianNetwork.domain)) },
     { case seed => Region.eqRegionIterable(AlarmBurglaryEarthquakeBayesianNetwork.domain) }
 )(
     axle.pgm.MonotypeBayesanNetwork.probabilityModelForMonotypeBayesanNetwork[Boolean, DirectedSparseGraph],
